+++ httpcore/_sync/connection.py
@@ -172,6 +172,9 @@ class HTTPConnection(ConnectionInterface):
             with Trace("close", logger, None, {}):
                 self._connection.close()
 
+    def is_connected(self) -> bool:
+        return self._connection is not None and self._connection.is_connected()
+
     def is_available(self) -> bool:
         if self._connection is None:
             # If HTTP/2 support is enabled, and the resulting connection could