+++ httpcore/_sync/socks_proxy.py
@@ -305,6 +305,9 @@ class Socks5Connection(ConnectionInterface):
         if self._connection is not None:
             self._connection.close()
 
+    def is_connected(self) -> bool:
+        return self._connection is not None and self._connection.is_connected()
+
     def is_available(self) -> bool:
         if self._connection is None:  # pragma: nocover
             # If HTTP/2 support is enabled, and the resulting connection could