+++ tests/test_cancellations.py
@@ -1,4 +1,5 @@
 import typing
+from unittest.mock import patch
 
 import anyio
 import hpack
@@ -133,6 +134,29 @@ async def test_connection_pool_timeout_during_response():
         with anyio.move_on_after(0.01):
             await pool.request("GET", "http://example.com")
         assert not pool.connections
+
+
+@pytest.mark.anyio
+async def test_connection_pool_cancellation_during_waiting_for_connection():
+    """
+    A cancellation of ongoing request waiting for a connection should leave
+    the pool in a consistent state.
+
+    In this case, that means the connection will become closed, and no
+    longer remain in the pool.
+    """
+
+    async def wait_for_connection(self, *args, **kwargs):
+        await anyio.sleep(999)
+
+    with patch(
+        "httpcore._async.connection_pool.AsyncPoolRequest.wait_for_connection",
+        new=wait_for_connection,
+    ):
+        async with httpcore.AsyncConnectionPool() as pool:
+            with anyio.move_on_after(0.01):
+                await pool.request("GET", "http://example.com")
+            assert not pool.connections
 
 
 @pytest.mark.anyio